adding nuber2Words methods to Piglatin class
